feat: Add database models

+++ spacex-api-wrapper/src/main/java/com/haroldadmin/spacex_api_wrapper/common/Mass.kt
@@ -5,6 +5,6 @@ import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
 data class Mass(
-    @field:Json(name = "kg") val kg: Int,
-    @field:Json(name = "lb") val lb: Int
+    @field:Json(name = "kg") val kg: Double,
+    @field:Json(name = "lb") val lb: Double
 )
