+++ spacex-api-wrapper/src/main/java/com/haroldadmin/spacex_api_wrapper/common/Thrust.kt
@@ -6,5 +6,5 @@ import com.squareup.moshi.JsonClass
 @JsonClass(generateAdapter = true)
 data class Thrust(
     @field:Json(name = "kN") val kN: Double,
-    @field:Json(name = "lbf") val lbf: Int
+    @field:Json(name = "lbf") val lbf: Double
 )
